Rename freestanding to solo5
