+++ install.sh
@@ -5,9 +5,9 @@ if [ "$prefix" = "" ]; then
     prefix=`opam var prefix`
 fi
 
-DESTINC=${prefix}/freestanding-sysroot/include/nolibc
-DESTLIB=${prefix}/freestanding-sysroot/lib/nolibc
-SYSROOT=${prefix}/freestanding-sysroot
+DESTINC=${prefix}/solo5-sysroot/include/nolibc
+DESTLIB=${prefix}/solo5-sysroot/lib/nolibc
+SYSROOT=${prefix}/solo5-sysroot
 mkdir -p ${DESTINC} ${DESTLIB} ${SYSROOT}
 
 # nolibc
@@ -29,10 +29,10 @@ touch ${DESTLIB}/META
 
 # findlib
 mkdir -p ${prefix}/lib/findlib.conf.d
-cp freestanding.conf ${prefix}/lib/findlib.conf.d/freestanding.conf
+cp solo5.conf ${prefix}/lib/findlib.conf.d/solo5.conf
 
 # dummy packages
 mkdir -p ${SYSROOT}/lib/threads
 touch ${SYSROOT}/lib/threads/META # for ocamlfind
-mkdir -p ${SYSROOT}/lib/is_freestanding
-touch ${SYSROOT}/lib/is_freestanding/META
+mkdir -p ${SYSROOT}/lib/is_solo5
+touch ${SYSROOT}/lib/is_solo5/META
