+++ nolibc/include/_solo5/byteorder.h
