+++ nolibc/include/sys/param.h
@@ -1,6 +1,6 @@
 #ifndef _SYS_PARAM_H
 #define _SYS_PARAM_H
 
-#include <_freestanding/byteorder.h>
+#include <_solo5/byteorder.h>
 
 #endif /* _SYS_PARAM_H */
