+++ ocaml-solo5.opam
@@ -1,15 +1,16 @@
 opam-version: "2.0"
 maintainer: "Martin Lucina <martin@lucina.net>"
 authors: "Martin Lucina <martin@lucina.net>"
-homepage: "https://github.com/mirage/ocaml-freestanding"
-bug-reports: "https://github.com/mirage/ocaml-freestanding/issues/"
+homepage: "https://github.com/mirage/ocaml-solo5"
+bug-reports: "https://github.com/mirage/ocaml-solo5/issues/"
 license: "MIT"
 tags: "org:mirage"
-dev-repo: "git+https://github.com/mirage/ocaml-freestanding.git"
+dev-repo: "git+https://github.com/mirage/ocaml-solo5.git"
 build: [
   ["./configure.sh"
    "--prefix=%{prefix}%"
-   "--target=aarch64-solo5-none-static"
+   "--target=x86_64-solo5-none-static" { arch = "x86_64" }
+   "--target=aarch64-solo5-none-static" { arch = "arm64" }
    "--ocaml-configure-option=--disable-flat-float-array" {ocaml-option-no-flat-float-array:installed}
    "--ocaml-configure-option=--enable-flambda" {ocaml-option-flambda:installed}
    "--ocaml-configure-option=--disable-naked-pointers" {ocaml-option-nnp:installed}
@@ -23,10 +24,8 @@ depends: [
   "ocaml-src" {build}
   "ocaml" {>= "4.12.1" & < "4.15.0"}
   "solo5" {>= "0.7.0"}
-  "solo5-cross-aarch64" {>= "0.7.0" }
 ]
 conflicts: [
-  "ocaml-freestanding"
   "sexplib" {= "v0.9.0"}
   "solo5-kernel-ukvm"
   "solo5-kernel-virtio"
@@ -39,5 +38,5 @@ available: [
 ]
 synopsis: "Freestanding OCaml compiler"
 description:
-  "This package provides a freestanding OCaml cross-compiler for ARM64, suitable for linking with a unikernel base layer."
+  "This package provides a OCaml cross-compiler, suitable for linking with a Solo5 unikernel."
 version: "0.7.0"
